requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
joblib==1.1.0
numpy==1.22.4
pyspark==3.1.2
pytz==2022.1
Sastrawi==1.0.1
scikit-learn==1.1.1
scipy==1.8.1
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
